initial commit from docker

+++ src/observer/common/value.cpp
@@ -493,7 +493,7 @@ std::vector<float> Value::get_vector() const
 {
   assert(attr_type_ == AttrType::VECTORS);
   std::vector<float> vector(get_vector_length());
-  for (int i = 0; i < vector.size(); ++i) {
+  for (size_t i = 0; i < vector.size(); ++i) {
     vector[i] = get_vector_element(i);
   }
   return vector;
